# contracts/openapi/core-api.yaml
openapi: 3.1.0
x-copyright: Copyright © kexyz254peter. Nexuss AI - Confidential and Proprietary.
info:
  title: Nexuss Core API
  version: 0.5.1
  description: Confidential P5.1 intent-aware media, paired-phone, and trusted-device contract - kexyz254peter
paths:
  /health/live:
    get:
      operationId: healthLive
      responses:
        '200':
          description: Process is alive
          content:
            application/json:
              schema:
                type: object
                required: [status, service, version]
                properties:
                  status: {type: string, const: alive}
                  service: {type: string, const: nexuss-core-api}
                  version: {type: string, const: 0.5.1}
  /health/ready:
    get:
      operationId: healthReady
      responses:
        '200':
          description: P5.1 UI, media provider, paired-phone gateway, and trusted-device client are ready
          content:
            application/json:
              schema:
                type: object
                required: [status, mode, phone_approval]
                properties:
                  status: {type: string, const: ready}
                  mode: {type: string, const: p5_knowledge_media_mobile}
                  phone_approval: {type: string, const: enabled}
  /v1/capabilities:
    get:
      operationId: listCapabilities
      responses:
        '200': {description: Explicitly registered capability manifests}
  /v1/tasks:
    post:
      operationId: createTask
      parameters:
        - in: header
          name: X-Nexuss-Session-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Session-Authenticated
          required: true
          schema: {type: boolean}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/task-request.schema.json'
      responses:
        '202': {description: Task interpreted, planned, and policy evaluated}
        '401': {description: Invalid user session}
  /v1/tasks/{task_id}:
    get:
      operationId: getTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Current task state}
        '404': {description: Task not found}
  /v1/tasks/{task_id}/approval:
    post:
      operationId: decideDesktopTaskApproval
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Session-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Session-Authenticated
          required: true
          schema: {type: boolean}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '../schemas/approval-decision.schema.json'
      responses:
        '200': {description: Desktop approval or cancellation applied}
        '401': {description: Session mismatch}
        '404': {description: Task not found}
        '409': {description: Phone approval required or approval invalid}
  /v1/mobile/pairing:
    post:
      operationId: createPhonePairingChallenge
      parameters:
        - in: header
          name: X-Nexuss-Session-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Session-Authenticated
          required: true
          schema: {type: boolean}
      responses:
        '200': {description: Single-use session-bound eight-digit pairing challenge}
        '401': {description: Desktop session is not authenticated}
  /v1/mobile/pair:
    post:
      operationId: pairPhone
      responses:
        '200': {description: Paired phone bearer session}
        '409': {description: Pairing invalid, expired, or already consumed}
  /v1/mobile/pending:
    get:
      operationId: listPhoneApprovals
      parameters:
        - in: header
          name: X-Nexuss-Mobile-Device-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Mobile-Token
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Pending approvals bound to the paired desktop session
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../schemas/mobile-approval.schema.json'
        '401': {description: Phone session invalid or expired}
  /v1/mobile/handoffs:
    get:
      operationId: claimPairedPhoneHandoffs
      parameters:
        - in: header
          name: X-Nexuss-Mobile-Device-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Mobile-Token
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Once-only allowlisted YouTube handoffs for the paired phone
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '../schemas/mobile-handoff.schema.json'
        '401': {description: Phone session invalid or expired}
  /v1/mobile/tasks/{task_id}/decision:
    post:
      operationId: decidePhoneApproval
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Mobile-Device-ID
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Mobile-Token
          required: true
          schema: {type: string}
      responses:
        '200': {description: Phone decision applied and device command lifecycle advanced}
        '401': {description: Phone session invalid or expired}
        '404': {description: Task not found}
        '409': {description: Approval invalid, expired, modified, or consumed}
  /v1/tasks/{task_id}/rollback:
    post:
      operationId: rollbackTask
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
        - in: header
          name: X-Nexuss-Session-ID
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Receipt-bound file or device-process rollback completed}
        '401': {description: Session mismatch}
        '404': {description: Task not found}
        '409': {description: Rollback unavailable or verification failed}
  /v1/tasks/{task_id}/receipt:
    get:
      operationId: getTaskReceipt
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200':
          description: Latest immutable Action Receipt snapshot
          content:
            application/json:
              schema:
                $ref: '../schemas/action-receipt.schema.json'
        '404': {description: Receipt not found}
  /v1/tasks/{task_id}/receipts:
    get:
      operationId: getTaskReceiptHistory
      parameters:
        - in: path
          name: task_id
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Append-only receipt history ordered by receipt version}
        '404': {description: Receipt history not found}